feat : 커스텀 에러 작성
