+++ src/main/java/numble/instagram/common/exception/ConflictException.java
@@ -1,0 +1,8 @@
+package numble.instagram.common.exception;
+
+public class ConflictException extends RuntimeException {
+
+    public ConflictException(String message) {
+        super(message);
+    }
+}
